Добавлено: 1. Обработка исключений 2. Группировка товаров в корзине с подсчетом сумм по каждой позиции и итоговой суммой 3. Страница формирования заказа с выводом информации о клиенте, выводом списка товаров с подсчетом сумм по каждой позиции и итоговой суммой и выбором способа доставки 4. Созранение заказа в базе данных (не самая изящная реализация, но работает)

+++ src/main/java/ru/gb/data/Product.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 
 @Data
 @NoArgsConstructor
@@ -19,9 +21,9 @@ public class Product {
     @Column(name = "name")
     private String name;
     @Column(name = "cost")
-    private Double cost;
+    private BigDecimal cost;
 
-    public Product(final String name, final Double cost) {
+    public Product(final String name, final BigDecimal cost) {
         this.name = name;
         this.cost = cost;
     }
